server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0  
apscheduler==3.10.1 
requests==2.31.0
python-dotenv==1.0.0